Feature/sft revision handling (#136)
* [ SFT REVISION ]minor, FILE BSP API attribut type change

* [ SFT REVISION ] two new BSP FWU functions: get and set revision. Implementation and invocation of these functions from SFT logic. A POSIX implementation is added which creates an extra file with “.xirev” suffix (concatenated to the Xively filename). The content of the file is the revision received through SFT. The POSIX implementation uses “only” BSP functions so in theory this isn’t POSIX but cross-platform. Next step will be to test this on CC3200.

* [ SFT REVISION ] providing BSP FWU for the dummy layer, ARM target uses dummy BSP

* [ SFT REVISION ] adding common implementation for both POSIX and CC3200 platforms. It saves revision in an extra file, suffixed with .xirev.

* [ SFT REVISION ] Travis bulild fix: providing “not implemented” bodies for BSP FILE function for dummy BSP module

* [ STM REVISION ] ARM target link fix: addding unresolved symbol

* [ SFT REVISION ] self review

* [ SFT REVISION ] since the revision handling was moved back to inside libxively the internal memory management can be used which is tracked == leak detection is applied on it

* [ SFT REVISION ] further memory handling simplification could be done due revision is used directly from inside libxively

* [ SFT REVISION ] deletion of .xirev files generated by integration tests

* [ SFT REVISION ] review change

+++ src/bsp/platform/cc3200/xi_bsp_io_fs_cc3200.c
@@ -82,7 +82,7 @@ _u32 xi_bsp_io_fs_open_flags_to_sl_flags( const uint32_t size,
 }
 
 xi_state_t xi_bsp_io_fs_open( const char* const resource_name,
-                              const uint32_t size,
+                              const size_t size,
                               const xi_fs_open_flags_t open_flags,
                               xi_fs_resource_handle_t* resource_handle_out )
 {
